* dev: (37 commits)
  misc colors qa fixes
  Add missing semicolon
  Add missing semicolon
  bigint stuff
  Return all necessary fields for proposal from all the endpoints
  server error colors qa fixes
  more settings colors qa fixes
  settings color qa fixes
  Cleanup SqlCache and remove Jedis dependency
  props and project details
  small settings color qa fixes
  drag handle color qa fixes
  project right panel color qa fixes
  project left panel color qa fixes
  props and project details
  work queue colors
  update success snackbar color
  Contacts and Users color qa fixes
  fix
  commissions color fixes qa AND all the No data available messages (hopefully)
  ...

+++ src/main/resources/db/migration/procedures/brs/commissions/R__get_commission_account_details_for_setters.sql
@@ -1,15 +1,15 @@
-DROP FUNCTION IF EXISTS brs.get_commission_account_details_for_setters(integer,BIGINT [],INTEGER,INTEGER,DATE,DATE,INTEGER,INTEGER);
-CREATE OR REPLACE FUNCTION brs.get_commission_account_details_for_setters(p_payroll_id         integer,
+DROP FUNCTION IF EXISTS brs.get_commission_account_details_for_setters(bigint,BIGINT [],bigint,bigint,DATE,DATE,bigint,bigint);
+CREATE OR REPLACE FUNCTION brs.get_commission_account_details_for_setters(p_payroll_id         bigint,
                                                               p_project_ids         BIGINT [],
-                                                              p_contact_id   INTEGER,
-                                                              p_sales_rep          INTEGER,
+                                                              p_contact_id   bigint,
+                                                              p_sales_rep          bigint,
                                                               p_cancel_start_date  DATE,
                                                               p_cancel_end_date    DATE,
-                                                              p_override_plan_id   INTEGER,
-                                                              p_commission_plan_id INTEGER)
+                                                              p_override_plan_id   bigint,
+                                                              p_commission_plan_id bigint)
     RETURNS TABLE(
-                     project_id                                  INTEGER,
-                     customer_id                                 INT,
+                     project_id                                  bigint,
+                     customer_id                                 bigint,
                      project_name                               VARCHAR,
                      user_id                              bigint,
                      setter                                      TEXT,
@@ -20,9 +20,9 @@ CREATE OR REPLACE FUNCTION brs.get_commission_account_details_for_setters(p_payr
                      closer_appointment_outcome                         varchar,
                      overrides_per_user                          JSON,
                      override_plan                               TEXT,
-                     override_plan_id                            INT,
+                     override_plan_id                            bigint,
                      commission_plan                             TEXT,
-                     commission_plan_id                          INT,
+                     commission_plan_id                          bigint,
                      total_commissions                           NUMERIC(10,2),
                      total_overrides                             NUMERIC(10,2),
                      commission_earned                           NUMERIC(10,2),
@@ -174,7 +174,7 @@ BEGIN
                         0::numeric AS override_adjustments,
                         (SELECT coalesce(sum(amount), 0)
                          FROM brs.project_commission_ledger dcl
-                         WHERE dcl.project_id = p.id::integer
+                         WHERE dcl.project_id = p.id::bigint
                            AND dcl.ledger_type_id = 1
                            and dcl.position_id = 4)
                             AS commission_paid_to_date,
